Step 1 of expiring dates: Receive the data from the user.
Created an edit text that shows a DatePickerFragment and lets the user pick a date. Once they click ok, the fragment closes and the edit text receives the date and puts it in format.

Next: Handle the information received and put in the database. Then, add a textView to show it in every post. Finally, add an option to show only valid coupons (not expired ones).